// sdks/typescript/src/apache_beam/proto/beam_artifact_api.client.ts
export interface IArtifactRetrievalServiceClient {
    /**
     * Resolves the given artifact references into one or more replacement
     * artifact references (e.g. a Maven dependency into a (transitive) set
     * of jars.
     *
     * @generated from protobuf rpc: ResolveArtifacts(org.apache.beam.model.job_management.v1.ResolveArtifactsRequest) returns (org.apache.beam.model.job_management.v1.ResolveArtifactsResponse);
     */
    resolveArtifacts(input: ResolveArtifactsRequest, options?: RpcOptions): UnaryCall<ResolveArtifactsRequest, ResolveArtifactsResponse>;
    /**
     * Retrieves the given artifact as a stream of bytes.
     *
     * @generated from protobuf rpc: GetArtifact(org.apache.beam.model.job_management.v1.GetArtifactRequest) returns (stream org.apache.beam.model.job_management.v1.GetArtifactResponse);
     */
    getArtifact(input: GetArtifactRequest, options?: RpcOptions): ServerStreamingCall<GetArtifactRequest, GetArtifactResponse>;
}
/**
 * A service to retrieve artifacts for use in a Job.
 *
 * @generated from protobuf service org.apache.beam.model.job_management.v1.ArtifactRetrievalService
 */
export class ArtifactRetrievalServiceClient implements IArtifactRetrievalServiceClient, ServiceInfo {
    typeName = ArtifactRetrievalService.typeName;
    methods = ArtifactRetrievalService.methods;
    options = ArtifactRetrievalService.options;
    constructor(private readonly _transport: RpcTransport) {
    }
    /**
     * Resolves the given artifact references into one or more replacement
     * artifact references (e.g. a Maven dependency into a (transitive) set
     * of jars.
     *
     * @generated from protobuf rpc: ResolveArtifacts(org.apache.beam.model.job_management.v1.ResolveArtifactsRequest) returns (org.apache.beam.model.job_management.v1.ResolveArtifactsResponse);
     */
    resolveArtifacts(input: ResolveArtifactsRequest, options?: RpcOptions): UnaryCall<ResolveArtifactsRequest, ResolveArtifactsResponse> {
        const method = this.methods[0], opt = this._transport.mergeOptions(options);
        return stackIntercept<ResolveArtifactsRequest, ResolveArtifactsResponse>("unary", this._transport, method, opt, input);
    }
    /**
     * Retrieves the given artifact as a stream of bytes.
     *
     * @generated from protobuf rpc: GetArtifact(org.apache.beam.model.job_management.v1.GetArtifactRequest) returns (stream org.apache.beam.model.job_management.v1.GetArtifactResponse);
     */
    getArtifact(input: GetArtifactRequest, options?: RpcOptions): ServerStreamingCall<GetArtifactRequest, GetArtifactResponse> {
        const method = this.methods[1], opt = this._transport.mergeOptions(options);
        return stackIntercept<GetArtifactRequest, GetArtifactResponse>("serverStreaming", this._transport, method, opt, input);
    }
}